requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# AI Agents Framework (Optional - commented out due to Python 3.13 compatibility)
# crewai==0.1.27
# langchain==0.1.0
# langchain-openai==0.0.2
# langgraph==0.0.20

# LLM Integration
openai==1.10.0
anthropic==0.8.1
tiktoken==0.5.2
google-generativeai==0.3.2

# Web Automation
selenium==4.16.0
playwright==1.41.0
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.3
lxml==5.1.0

# Vector Database & Embeddings
chromadb==0.4.22
pinecone-client==3.0.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# NLP & Text Processing (Optional - simplified for Python 3.13)
# spacy==3.7.2
nltk==3.8.1
# transformers==4.37.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1

# Data Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
cryptography>=40.0.0,<42.0.0
authlib==1.3.0

# Task Scheduling
celery==5.3.4
apscheduler==3.10.4

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.40.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3

# Frontend (Streamlit)
streamlit==1.30.0
plotly==5.18.0
pandas==2.1.4

# Code Quality
black==23.12.1
pylint==3.0.3
flake8==7.0.0
mypy==1.8.0
bandit==1.7.6

# Email
sendgrid==6.11.0
python-email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0
